# openapi.yaml
openapi: 3.0.0
info:
  title: Gol Ahora - API de Gestión
  description: |
    API de gestión y automatización de procesos de administración, alquiler de canchas y actividades del complejo deportivo **"El buen deporte"**.
    
    ### Niveles de Acceso y Roles:
    1. **Usuario (Público - Visitante no registrado)**: Puede registrarse, listar canchas/tipos, ver la disponibilidad de horarios y ver la oferta pública de clases y entrenamientos.
    2. **Cliente (Registrado)**: Deportista con cuenta. Puede gestionar su perfil, realizar, modificar y cancelar reservas (previo pago), e inscribirse o darse de baja en clases y entrenamientos grupales o particulares. Solo consulta su propia facturación (cobros/recibos). No participa directamente en la creación ni inscripción de ligas/torneos.
    3. **Administrador (Personal/Staff del club)**: Empleados y coordinadores. Tienen el control total de la gestión de clientes, profesores/entrenadores (certificaciones), canchas físicas (mantenimiento y horarios), políticas de descuentos, cobros globales, organización de ligas y torneos (inscripción de equipos, generación de fixtures, brackets y resultados de partidos) y generación de reportes financieros y de asistencia.
  version: 1.0.0
servers:
  - url: /
    description: Servidor principal (Autodetectado)
...
paths:
...
  /usuario/login:
    post:
      tags:
        - Gestión de Clientes (Público)
      summary: Iniciar sesión en el sistema (Clientes y Administradores)
      description: Permite a los clientes y al personal del club ingresar con sus credenciales y obtener su ID y nivel de acceso.
      operationId: iniciarSesion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginEntrada'
      responses:
        '200':
          description: Inicio de sesión exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  usuario:
                    $ref: '#/components/schemas/Cliente'
        '401':
          description: Credenciales inválidas (usuario/correo o contraseña incorrectos)
        '400':
          description: Faltan campos obligatorios
...
  /profesor/info:
    get:
      tags:
        - Gestión de Profesores
      summary: Consultar datos personales del profesor logueado
      description: Retorna la información de perfil del profesor autenticado con la dirección estructurada de forma anidada.
      operationId: obtenerInfoProfesor
      responses:
        '200':
          description: Datos del profesor logueado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoUsuario'
        '401':
          description: No autorizado
        '404':
          description: Profesor no encontrado
  /entrenador/info:
    get:
      tags:
        - Gestión de Entrenadores
      summary: Consultar datos personales del entrenador logueado
      description: Retorna la información de perfil del entrenador autenticado con la dirección estructurada de forma anidada.
      operationId: obtenerInfoEntrenador
      responses:
        '200':
          description: Datos del entrenador logueado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoUsuario'
        '401':
          description: No autorizado
        '404':
          description: Entrenador no encontrado

  # ==========================================
  # === 2. ACCESO CLIENTE (REGISTRADO) ===
  # ==========================================
...
components:
  schemas:
    InfoUsuario:
      type: object
      properties:
        id_usuario:
          type: integer
          example: 8
        username:
          type: string
          example: profe_carlos
        nombre:
          type: string
          example: Carlos
        apellido:
          type: string
          example: Gómez
        email:
          type: string
          example: carlos.gomez@golahora.com
        fecha_nacimiento:
          type: string
          format: date
          example: '1985-05-15'
        dni:
          type: string
          example: '30222111'
        telefono:
          type: string
          example: '1122334455'
        genero:
          type: string
          example: Masculino
        nacionalidad:
          type: string
          example: Argentina
        direccion:
          type: object
          properties:
            calle:
              type: string
              example: Avenida Calchaquí
            numero:
              type: string
              example: '6200'
            codigo_postal:
              type: string
              example: '1888'
            localidad:
              type: string
              example: Florencio Varela
            provincia:
              type: string
              example: Buenos Aires
            pais:
              type: string
              example: Argentina
    Cliente:
      type: object
      properties:
        id_usuario:
          type: integer
        username:
          type: string
        user_level:
          type: integer
        nombre:
          type: string
        apellido:
          type: string
        email:
          type: string
          format: email
        dni:
          type: string
        telefono:
          type: string
        fecha_nacimiento:
          type: string
          format: date
        fecha_registro:
          type: string
          format: date-time
        genero:
          type: string
        nacionalidad:
          type: string
        calle:
          type: string
        numero:
          type: string
        codigo_postal:
          type: string
        localidad:
          type: string
        provincia:
          type: string
      required:
        - id_usuario
        - username
        - user_level
        - nombre
        - apellido
        - email
        - dni
        - fecha_nacimiento
        - fecha_registro
        - genero
        - nacionalidad
        - calle
        - numero
        - codigo_postal
        - localidad
...
    LoginEntrada:
      type: object
      properties:
        username:
          type: string
          description: Nombre de usuario o correo electrónico de acceso
        password:
          type: string
          description: Contraseña del usuario
      required:
        - username
        - password
...